FFWEB-2023: Refactor of CustomFields.php
